Remove obselete fee cells
- Rename FeeCellNew to FeeCell
- Remove FeeSlider, EditableFeeCell
